change: refactor GroupController
