+++ src/main/java/me/universi/group/entities/GroupSettings/GroupFeatures.java
@@ -6,7 +6,9 @@ import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
 import java.io.Serial;
 import java.io.Serializable;
+import java.util.Date;
 import java.util.UUID;
+import org.hibernate.annotations.CreationTimestamp;
 import org.hibernate.annotations.SQLDelete;
 import org.hibernate.annotations.Where;
 
@@ -19,7 +21,6 @@ public class GroupFeatures  implements Serializable {
     @Serial
     private static final long serialVersionUID = -31637842245354343L;
 
-    @JsonIgnore
     @Id
     @GeneratedValue(strategy = GenerationType.UUID)
     @Column(name = "id")
@@ -32,20 +33,50 @@ public class GroupFeatures  implements Serializable {
 
     @JsonIgnore
     @JoinColumn(name="group_settings_id")
-    @OneToOne(fetch = FetchType.EAGER)
+    @ManyToOne(fetch = FetchType.EAGER)
     @NotNull
     public GroupSettings groupSettings;
 
-    @Column(name = "contents")
-    public boolean showContents = Boolean.TRUE;
+    @Column(name = "name")
+    @NotNull
+    public String name;
 
-    @Column(name = "groups")
-    public boolean showGroups = Boolean.TRUE;
+    @Column(name = "description")
+    public String description;
 
-    @Column(name = "participants")
-    public boolean showParticipants = Boolean.TRUE;
+    @Column(name = "enabled")
+    public boolean enabled = Boolean.TRUE;
+
+    @CreationTimestamp
+    @Column(name = "added")
+    @Temporal(TemporalType.TIMESTAMP)
+    public Date added;
+
+    @JsonIgnore
+    @Column(name = "removed")
+    @Temporal(TemporalType.TIMESTAMP)
+    public Date removed;
 
     public GroupFeatures() {
     }
 
+    public boolean isDeleted() { return deleted; }
+
+    public void setDeleted(boolean deleted) { this.deleted = deleted; }
+
+    public Date getRemoved() {
+        return removed;
+    }
+
+    public void setRemoved(Date removed) {
+        this.removed = removed;
+    }
+
+    public Date getAdded() {
+        return added;
+    }
+
+    public void setAdded(Date added) {
+        this.added = added;
+    }
 }
