+++ src/main/java/me/universi/group/services/GroupService.java
@@ -642,7 +642,47 @@ public class GroupService {
         return true;
     }
 
-    public boolean editFeatures(Group group, Boolean showContents, Boolean showGroups, Boolean showParticipants) {
+    public boolean addFeature(Group group, String name, String description, Boolean enabled) {
+        if(group == null || name == null || name.isEmpty()) {
+            return false;
+        }
+        GroupSettings groupSettings = group.getGroupSettings();
+        if(groupSettings == null) {
+            return false;
+        }
+        GroupFeatures groupFeature = new GroupFeatures();
+        groupFeature.groupSettings = groupSettings;
+        groupFeature.name = name;
+        groupFeature.description = description;
+        if(enabled != null) {
+            groupFeature.enabled = enabled;
+        }
+        groupFeaturesRepository.save(groupFeature);
+        return true;
+    }
+
+    public boolean deleteFeature(Group group, UUID groupFeatureId) {
+        if(group == null || groupFeatureId == null) {
+            return false;
+        }
+        if(groupFeaturesRepository.existsByGroupSettingsIdAndId(group.getGroupSettings().getId(), groupFeatureId)) {
+            GroupFeatures groupFeature = groupFeaturesRepository.findFirstByGroupSettingsIdAndId(group.getGroupSettings().getId(), groupFeatureId);
+            groupFeature.setRemoved(ConvertUtil.getDateTimeNow());
+            groupFeature.setDeleted(true);
+            groupFeaturesRepository.save(groupFeature);
+            return true;
+        }
+        return false;
+    }
+
+    public boolean deleteFeature(Group group, String groupFeatureId) {
+        if(group == null || groupFeatureId == null || groupFeatureId.isEmpty()) {
+            return false;
+        }
+        return deleteFeature(group, UUID.fromString(groupFeatureId));
+    }
+
+    public boolean editFeature(Group group, UUID groupFeatureId, Boolean enabled, String description) {
         if(group == null) {
             return false;
         }
@@ -650,23 +690,25 @@ public class GroupService {
         if(groupSettings == null) {
             return false;
         }
-        GroupFeatures groupFeatures = groupSettings.features;
-        if(groupFeatures == null) {
-            groupFeatures = new GroupFeatures();
-            groupFeatures.groupSettings = groupSettings;
-            groupFeatures = groupFeaturesRepository.save(groupFeatures);
+        GroupFeatures groupFeature = groupFeaturesRepository.findFirstByGroupSettingsIdAndId(groupSettings.getId(), groupFeatureId);
+        if(groupFeature == null) {
+            throw new GroupException("Feature não encontrada.");
         }
-        if(showContents != null) {
-            groupFeatures.showContents = showContents;
+        if(enabled != null) {
+            groupFeature.enabled = enabled;
         }
-        if(showGroups != null) {
-            groupFeatures.showGroups = showGroups;
+        if(description != null) {
+            groupFeature.description = description;
         }
-        if(showParticipants != null) {
-            groupFeatures.showParticipants = showParticipants;
-        }
-        groupFeaturesRepository.save(groupFeatures);
+        groupFeaturesRepository.save(groupFeature);
         return true;
+    }
+
+    public boolean editFeature(Group group, String groupFeatureId, Boolean enabled, String description) {
+        if(group == null || groupFeatureId == null || groupFeatureId.isEmpty()) {
+            return false;
+        }
+        return editFeature(group, UUID.fromString(groupFeatureId), enabled, description);
     }
 
     public boolean addAdministrator(Group group, Profile profile) {
